+++ R/auk-clean.r
@@ -10,9 +10,12 @@
 #' @param sep character; the input field separator, the basic dataset is tab
 #'   separated by default. Must only be a single character and space delimited
 #'   is not allowed since spaces appear in many of the fields.
-#' @param remove_blank logical; whether the trailing blank should be removed
-#'   from the end of each row. The EBD comes with an extra tab at the end of
-#'   each line, which causes a extra blank column.
+#' @param remove_text logical; whether all free text entry columns should be 
+#'   removed. These columns include comments, location names, and observer 
+#'   names. These columns cause import errors due to special characters and 
+#'   increase the file size, yet are rarely valuable for analytical 
+#'   applications, so may be removed. Setting this argument to `TRUE` can lead 
+#'   to a significant reduction in file size.
 #' @param overwrite logical; overwrite output file if it already exists
 #'
 #' @details
@@ -43,8 +46,8 @@
 #' ncol(read.delim(f, nrows = 5, quote = ""))
 #' ncol(read.delim(tmp, nrows = 5, quote = ""))
 #' }
-auk_clean <- function(f_in, f_out, sep = "\t",
-                      remove_blank = TRUE, overwrite = FALSE) {
+auk_clean <- function(f_in, f_out, sep = "\t", remove_text = FALSE, 
+                      overwrite = FALSE) {
   # checks
   awk_path <- auk_getpath()
   if (is.na(awk_path)) {
@@ -53,7 +56,7 @@ auk_clean <- function(f_in, f_out, sep = "\t",
   assertthat::assert_that(
     file.exists(f_in),
     assertthat::is.string(sep), nchar(sep) == 1, sep != " ",
-    assertthat::is.flag(remove_blank),
+    assertthat::is.flag(remove_text),
     assertthat::is.flag(overwrite)
   )
   # check output file
@@ -68,7 +71,7 @@ auk_clean <- function(f_in, f_out, sep = "\t",
   # determine number of columns
   # read header row
   header <- get_header(f_in, sep)
-  if (remove_blank && header[length(header)] == "") {
+  if (header[length(header)] == "") {
     header <- header[-length(header)]
   }
   ncols <- length(header)
@@ -78,15 +81,23 @@ auk_clean <- function(f_in, f_out, sep = "\t",
             ncols)
       )
   }
+  
+  # columns to drop
+  if (remove_text) {
+    text_cols <- c("locality", 
+                   "first name", "last name", 
+                   "trip comments", 
+                   "species comments")
+    keep_cols <- which(!tolower(header) %in% text_cols)
+    print_cols <- paste0("$", keep_cols, collapse = ",")
+  } else {
+    print_cols <- "$0"
+  }
+  
 
   # construct awk command
-  if (remove_blank) {
-    # remove end of line tab
-    ws <- "sub(/\t$/, \"\", $0)"
-  } else {
-    ws <- ""
-  }
-  awk <- str_interp(awk_clean, list(ncols = ncols, ws = ws, sep = sep))
+  awk <- str_interp(awk_clean, 
+                    list(sep = sep, ncols = ncols, print_cols = print_cols))
 
   # run command
   exit_code <- system2(awk_path,
@@ -106,10 +117,11 @@ BEGIN {
   OFS = \"${sep}\"
 }
 {
-  ${ws}
+  # remove end of line tab
+  sub(/\t$/, \"\", $0)
   # only keep rows with correct number of records
   if (NF == ${ncols} || NR == 1) {
-    print $0
+    print ${print_cols}
   }
 }
 "