+++ R/auk-complete.r
@@ -7,7 +7,11 @@
 #' filters have been defined, [auk_filter()] should be used to call AWK and
 #' perform the filtering.
 #'
-#' @param x `auk_ebd` object; reference to basic dataset file created by [auk_ebd()].
+#' @param x `auk_ebd` or `auk_sampling` object; reference to file created by 
+#'   [auk_ebd()] or [auk_sampling()].
+#' 
+#' @details This function can also work with on an `auk_sampling` object if the 
+#'   user only wishes to filter the sampling event data.
 #'
 #' @return An `auk_ebd` object.
 #' @export
@@ -24,4 +28,9 @@ auk_complete.auk_ebd <- function(x) {
   # define filter
   x$filters$complete <- TRUE
   return(x)
+}
+
+#' @export
+auk_complete.auk_sampling <- function(x) {
+  auk_complete.auk_ebd(x)
 }