+++ R/auk-project.r
@@ -6,9 +6,13 @@
 #' once all filters have been defined, [auk_filter()] should be used to call AWK
 #' and perform the filtering.
 #'
-#' @param x `auk_ebd` object; reference to object created by [auk_ebd()].
+#' @param x `auk_ebd` or `auk_sampling` object; reference to file created by 
+#'   [auk_ebd()] or [auk_sampling()].
 #' @param project character; project code to filter by (e.g. `"EBIRD_MEX"`).
 #'   Multiple codes are accepted.
+#' 
+#' @details This function can also work with on an `auk_sampling` object if the 
+#'   user only wishes to filter the sampling event data.
 #'
 #' @return An `auk_ebd` object.
 #' @export
@@ -44,4 +48,9 @@ auk_project.auk_ebd <- function(x, project) {
   # set filter list
   x$filters$project <- project
   return(x)
+}
+
+#' @export
+auk_project.auk_sampling <- function(x, project) {
+  auk_project.auk_ebd(x, project)
 }