+++ R/auk-protocol.r
@@ -5,21 +5,25 @@
 #' once all filters have been defined, [auk_filter()] should be used to call AWK
 #' and perform the filtering.
 #'
-#' @param x `auk_ebd` object; reference to object created by [auk_ebd()].
-#' @param protocol character; "stationary", "traveling", or "causal". Other 
+#' @param x `auk_ebd` or `auk_sampling` object; reference to file created by 
+#'   [auk_ebd()] or [auk_sampling()].
+#' @param protocol character; "Stationary", "Traveling", or "Incidental". Other 
 #'   protocols exist in the database, however, this function only extracts these 
 #'   three standard protocols.
+#' 
+#' @details This function can also work with on an `auk_sampling` object if the 
+#'   user only wishes to filter the sampling event data.
 #'
 #' @return An `auk_ebd` object.
 #' @export
 #' @examples
 #' system.file("extdata/ebd-sample.txt", package = "auk") %>%
 #'   auk_ebd() %>%
-#'   auk_protocol("stationary")
+#'   auk_protocol("Stationary")
 #'   
 #' # alternatively, without pipes
 #' ebd <- auk_ebd(system.file("extdata/ebd-sample.txt", package = "auk"))
-#' auk_protocol(ebd, "stationary")
+#' auk_protocol(ebd, "Stationary")
 auk_protocol <- function(x, protocol)  {
   UseMethod("auk_protocol")
 }
@@ -27,10 +31,15 @@ auk_protocol <- function(x, protocol)  {
 #' @export
 auk_protocol.auk_ebd <- function(x, protocol) {
   assertthat::assert_that(
-    all(protocol %in% c("stationary", "traveling", "casual"))
+    all(protocol %in% c("Stationary", "Traveling", "Incidental"))
   )
   
   # set filter list
   x$filters$protocol <- protocol
   return(x)
+}
+
+#' @export
+auk_protocol.auk_sampling <- function(x, protocol) {
+  auk_protocol.auk_ebd(x, protocol)
 }